LC-maximum-total-subarray-value-i [Runtime: N/A; Memory: N/A]

+++ array/maximum-total-subarray-value-i.java
@@ -1,9 +1,9 @@
 class Solution {
     public long maxTotalValue(int[] nums, int k) {
         int i=0;
-        int min = Arrays.stream(nums).min().getAsInt();
-        int max = Arrays.stream(nums).max().getAsInt();
-       return (max-min)*k;
+        long min = Arrays.stream(nums).min().getAsInt();
+        long max = Arrays.stream(nums).max().getAsInt();
+       return (long)(max-min)*k;
         
     }
 }
